Added categories, fixing cart and so on

+++ cancel_item.php
@@ -8,22 +8,17 @@ if (!isset($_SESSION['loggedin'])) {
     exit();
 }
 
-if ($_SERVER['REQUEST_METHOD'] == 'POST') {
-    // Get the cart_id to be deleted
+if ($_SERVER['REQUEST_METHOD'] == 'POST' && isset($_POST['cart_id'])) {
     $cart_id = $_POST['cart_id'];
-
-    // Prepare the SQL to delete the item
     $sql = "DELETE FROM cart WHERE cart_id = ?";
-    if ($stmt = $conn->prepare($sql)) {
-        $stmt->bind_param("i", $cart_id);
-        if ($stmt->execute()) {
-            // Redirect to cart page with a success message
-            header('Location: cart.php?message=Item+successfully+canceled');
-            exit();
-        } else {
-            // Handle error
-            echo "Error: " . $stmt->error;
-        }
+    $stmt = $conn->prepare($sql);
+    $stmt->bind_param("i", $cart_id);
+    
+    if ($stmt->execute()) {
+        header('Location: cart.php?message=Item successfully cancelled');
+    } else {
+        header('Location: cart.php?message=Error cancelling item');
     }
+    exit();
 }
 ?>
